+++ dl4j-labs/src/main/java/global/skymind/solution/datavec/loadCSV/Exercise1.java
@@ -20,7 +20,7 @@
  *
  */
 
-package global.skymind.solution.datavec;
+package global.skymind.solution.datavec.loadCSV;
 
 import org.apache.commons.io.FileUtils;
 import org.datavec.api.records.reader.RecordReader;
