+++ dl4j-labs/src/main/java/global/skymind/solution/datavec/loadCSV/LoadCSV.java
@@ -20,7 +20,7 @@
  *
  */
 
-package global.skymind.solution.datavec;
+package global.skymind.solution.datavec.loadCSV;
 
 import org.datavec.api.records.reader.RecordReader;
 import org.datavec.api.records.reader.impl.csv.CSVRecordReader;
