+++ dl4j-labs/src/main/java/global/skymind/training/datavec/kfold/Exercise3.java
@@ -1,4 +1,4 @@
-package global.skymind.training.datavec;
+package global.skymind.training.datavec.kfold;
 
 import java.io.IOException;
