+++ dl4j-labs/src/main/java/global/skymind/training/datavec/loadCSV/LoadCSV.java
@@ -20,7 +20,7 @@
  *
  */
 
-package global.skymind.training.datavec;
+package global.skymind.training.datavec.loadCSV;
 
 import org.datavec.api.records.reader.RecordReader;
 import org.datavec.api.records.reader.impl.csv.CSVRecordReader;
