+++ dl4j-labs/src/main/java/global/skymind/training/datavec/loadImage/LoadImageAndTransform.java
@@ -20,7 +20,7 @@
  *
  */
 
-package global.skymind.solution.datavec;
+package global.skymind.training.datavec.loadImage;
 
 import org.datavec.api.io.filters.BalancedPathFilter;
 import org.datavec.api.io.labels.ParentPathLabelGenerator;
